add more tests and move to it's own file
