+++ pkg/services/ngalert/store/alert_rule_test.go
@@ -2,7 +2,6 @@ package store
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"strings"
@@ -538,140 +537,4 @@ func setupFolderService(t *testing.T, sqlStore *sqlstore.SQLStore, cfg *setting.
 	_, dashboardStore := testutil.SetupDashboardService(t, sqlStore, folderStore, cfg)
 
 	return testutil.SetupFolderService(t, cfg, dashboardStore, folderStore, inProcBus)
-}
-
-func TestMigrateLokiQueryToInstant(t *testing.T) {
-	original := &models.AlertRule{
-		Data: []models.AlertQuery{
-			{
-				RefID:         "A",
-				QueryType:     "range",
-				DatasourceUID: grafanaCloudLogs,
-				Model: []byte(`{
-					"datasource": {
-						"type": "loki",
-						"uid": "grafanacloud-logs"
-					},
-					"editorMode": "code",
-					"expr": "1",
-					"hide": false,
-					"intervalMs": 1000,
-					"maxDataPoints": 43200,
-					"queryType": "range",
-					"refId": "A"
-				}`),
-			},
-			{
-				RefID:         "B",
-				DatasourceUID: "__expr__",
-				Model: []byte(`{
-					"conditions": [
-						{
-							"evaluator": {
-								"params": [],
-								"type": "gt"
-							},
-							"operator": {
-								"type": "and"
-							},
-							"query": {
-								"params": [
-									"B"
-								]
-							},
-							"reducer": {
-								"params": [],
-								"type": "last"
-							},
-							"type": "query"
-						}
-					],
-					"datasource": {
-						"type": "__expr__",
-						"uid": "__expr__"
-					},
-					"expression": "A",
-					"hide": false,
-					"intervalMs": 1000,
-					"maxDataPoints": 43200,
-					"reducer": "last",
-					"refId": "B",
-					"type": "reduce"
-				}`),
-			},
-		},
-	}
-	mirgrated := &models.AlertRule{
-		Data: []models.AlertQuery{
-			{
-				RefID:         "A",
-				QueryType:     "instant",
-				DatasourceUID: grafanaCloudLogs,
-				Model: []byte(`{
-					"datasource": {
-						"type": "loki",
-						"uid": "grafanacloud-logs"
-					},
-					"editorMode": "code",
-					"expr": "1",
-					"hide": false,
-					"intervalMs": 1000,
-					"maxDataPoints": 43200,
-					"queryType": "instant",
-					"refId": "A"
-				}`),
-			},
-			{
-				RefID:         "B",
-				DatasourceUID: "__expr__",
-				Model: []byte(`{
-					"conditions": [
-						{
-							"evaluator": {
-								"params": [],
-								"type": "gt"
-							},
-							"operator": {
-								"type": "and"
-							},
-							"query": {
-								"params": [
-									"B"
-								]
-							},
-							"reducer": {
-								"params": [],
-								"type": "last"
-							},
-							"type": "query"
-						}
-					],
-					"datasource": {
-						"type": "__expr__",
-						"uid": "__expr__"
-					},
-					"expression": "A",
-					"hide": false,
-					"intervalMs": 1000,
-					"maxDataPoints": 43200,
-					"reducer": "last",
-					"refId": "B",
-					"type": "reduce"
-				}`),
-			},
-		},
-	}
-
-	require.True(t, canBeInstant(original))
-
-	migrateToInstant(original)
-
-	require.Equal(t, mirgrated.Data[0].QueryType, original.Data[0].QueryType)
-
-	originalModel := make(map[string]interface{})
-	require.NoError(t, json.Unmarshal(original.Data[0].Model, &originalModel))
-	migratedModel := make(map[string]interface{})
-	require.NoError(t, json.Unmarshal(mirgrated.Data[0].Model, &migratedModel))
-
-	require.Equal(t, migratedModel, originalModel)
 }
